Fixing quotes that span more than one line.
To break out of a quote-block in markdown, one has to have two line-breaks. It's the same as usual paragraphs, so having only one line-break results in it being collapsed as a single space. To represent that behavior here, I'm using start-end, with the previous match being the "start", and a blank line (^$) being the "end".